funcionalidad de obtencion de servicios

+++ entity/servicio.entity.go
@@ -8,5 +8,5 @@ type Service struct {
 	ID      uint   `gorm:"primaryKey;autoIncrement;column:id" json:"id"`
 	Nombre  string `gorm:"type:varchar(100);column:nombre" json:"nombre"`
 	Prefijo string `gorm:"type:varchar;not null;column:prefijo" json:"prefijo"`
-	Puerto  uint8  `gorm:"type:uint8;column:puerto" json:"puerto"`
+	Puerto  uint64 `gorm:"type:uint8;column:puerto" json:"puerto"`
 }
